refactor(botonic-react): create folder replies and styles file
