+++ packages/botonic-react/src/webchat/replies/index.tsx
@@ -1,36 +1,9 @@
 import React, { useContext } from 'react'
-import styled from 'styled-components'
 
-import { WEBCHAT } from '../constants'
-import { WebchatContext } from '../contexts'
-import { BotonicContainerId } from './constants'
-
-const ScrollableReplies = styled.div`
-  overscroll-behavior: contain;
-  -webkit-overflow-scrolling: touch;
-`
-
-interface RepliesContainerProps {
-  justify: string
-  wrap: string
-}
-
-const RepliesContainer = styled.div<RepliesContainerProps>`
-  display: flex;
-  text-align: center;
-  justify-content: ${props => props.justify};
-  flex-wrap: ${props => props.wrap};
-  padding-bottom: 10px;
-  margin-left: 5px;
-  margin-right: 5px;
-  overflow-x: auto;
-`
-
-const ReplyContainer = styled.div`
-  flex: none;
-  display: inline-block;
-  margin: 3px;
-`
+import { WEBCHAT } from '../../constants'
+import { WebchatContext } from '../../contexts'
+import { BotonicContainerId } from '../constants'
+import { RepliesContainer, ReplyContainer, ScrollableReplies } from './styles'
 
 const options = {
   left: 'flex-start',
